feat: card confirm 페이지에서 별칭 설정 - CardNickNameForm 분리
