+++ src/pages/CardRegisterConfirmPage.tsx
@@ -1,25 +1,41 @@
 import { useEffect, useState } from 'react';
 
-import { Button, CreditCard } from '@/components/UI';
-import { DefaultCardInfo } from '@/components/UI/CreditCard';
+import CardNickNameForm from '@/components/common/CardNickNameForm';
+import { Button, ColumnLayout, CreditCard, Text } from '@/components/UI';
 import { useRouter } from '@/hooks/useRouter';
 import { styled } from '@/lib/stitches.config';
 import { getItem } from '@/storage/storage';
 import { StorageKey } from '@/storage/storageKey';
+import { CardData } from '@/types';
 
 export const CardRegisterConfirmPage = () => {
-  const [card, setCard] = useState(DefaultCardInfo);
+  const [card, setCard] = useState<CardData>();
   const { go } = useRouter();
 
   useEffect(() => {
-    const newCard = getItem(StorageKey.CARD_LIST).at(-1);
-
+    const newCard = getItem(StorageKey.CARD_LIST).at(-1) as CardData;
     setCard(newCard);
   }, []);
+
+  if (!card) {
+    return (
+      <ColumnLayout css={{ height: '100%' }}>
+        <Text>카드가 존재하지 않습니다.</Text>
+        <Button
+          css={{ marginTop: '$3', width: '$10' }}
+          onClick={() => go('/list')}
+        >
+          카드 목록
+        </Button>
+      </ColumnLayout>
+    );
+  }
+
   return (
     <CardRegisterConfirmPageLayout>
-      <h2>카드 등록이 완료되었습니다.</h2>
       <CreditCard size="small" cardInfo={card} />
+      <h2>카드 등록이 완료되었습니다.</h2>
+      <CardNickNameForm card={card} />
       <Button
         css={{ position: 'absolute', bottom: '$5', width: '$11' }}
         onClick={() => go('/list')}
